refactor(docs): extract keyboard shortcut hook, use refs for stability
- Extract useVersionHistoryShortcut to packages/docs/src/hooks/
- Use refs for sidebarOpen/sidebarTab to register listener once
- Use snapshotsRef in handleRestore to avoid callback recreation
