+++ apps/web/src/features/docs/DocsEditorPage.tsx
@@ -9,6 +9,7 @@ import {
   useDocumentChat,
   BlockNoteEditor as BlockNoteEditorComponent,
   VersionHistory,
+  useVersionHistoryShortcut,
   useDocsAdapter,
   createDocsApiClient,
   lazyWithRetry,
@@ -332,21 +333,7 @@ function EditorContent() {
     setSidebarOpen((prev) => !prev);
   }, []);
 
-  useEffect(() => {
-    const handler = (e: KeyboardEvent) => {
-      if ((e.ctrlKey || e.metaKey) && e.shiftKey && e.key === 'h') {
-        e.preventDefault();
-        if (sidebarOpen && sidebarTab === 'versions') {
-          setSidebarOpen(false);
-        } else {
-          setSidebarTab('versions');
-          setSidebarOpen(true);
-        }
-      }
-    };
-    document.addEventListener('keydown', handler);
-    return () => document.removeEventListener('keydown', handler);
-  }, [sidebarOpen, sidebarTab]);
+  useVersionHistoryShortcut(sidebarOpen, sidebarTab, setSidebarOpen, setSidebarTab);
 
   useEffect(() => {
     setCommentsPortalTarget(
